requirements.txt
djangorestframework
openai
